Add program termination via button press
The program will terminate if the button is pressed a second time.

Also, I got rid of the pin toggle nomenclature for turning the LED on
and off. Applying a 1 to the LED pin in its pin vector would toggle the
pin according to Atmel datasheet section 14.2.2, but this syntax is
confusing. I changed it to the usual "turn this pin HI/LOW" convention.

I fixed a bug the button data direction register setting. This time it
will only place a 0 in the button's pin position on the DDR.