Fix gps indicator bug
The code was supposed to set a status variable when the gps fix
indicator in the GPGGA sentence indicated a good fix. The code was
looking for an indicator of 1. But, after looking at the raw bytes
in the datafiles, I saw that it was using the differential gps fix
indicator (2). I updated the code to allow this condition.

Hopefully this will fix the issue where the waypoint value is not
initially set.