Fixed gps signed coord bug
The nav-related gps values weren't being negated for the W/S
hemisphere conditions. This should fix the issue where the
zeroed-whole-degrees based nav coordinates were being calulated
in the wrong direction. For example, while the robot was moving
South, the nav breadcrumbs would be calculated as if the robot
was moving North.