Fix horrendous GPS headings
Instead of using the ground_course_deg value from the GPRMC sentence, I
created a function that will calculate the heading based on the previous
and current GPS coordinates.

Doing this meant that I could functionalize the calc_true_bearing code.
This function is used to calculate (1) the GPS heading and (2) the
relative bearing. As a result, it is called twice. Hopefully the main
loop doesn't run late more frequently.