Add mobility_init to startup
I also fixed a minor bug, which surprisingly wasn't caught during
compile.

The code was run twice and the initial results didn't look too good.
The robot drove off at almost full speed in a long left-hand turn.
I suspect that the PWM signal for the ESC isn't being terminated
in time. I have a fix in mind for this.